+++ packages/eslint-config-wesm87-base/rules/best-practices.js
@@ -195,6 +195,9 @@ module.exports = {
     // disallow use of assignment in return statement
     'no-return-assign': 'error',
 
+    // disallow use of await in return statement
+    'no-return-await': 'error',
+
     // disallow use of `javascript:` urls.
     'no-script-url': 'error',