+++ packages/eslint-config-wesm87/index.js
@@ -1,7 +1,7 @@
 module.exports = {
   extends: [
-    'eslint-config-airbnb-base',
-    'eslint-config-airbnb-base/rules/strict',
+    'eslint-config-wesm87-base',
+    'eslint-config-wesm87-base/rules/strict',
     './rules/react',
     './rules/react-a11y',
   ].map(require.resolve),