+++ packages/eslint-config-wesm87/legacy.js
@@ -1,0 +1,4 @@
+module.exports = {
+  extends: ['eslint-config-wesm87-base/legacy'].map(require.resolve),
+  rules: {},
+};