+++ packages/eslint-config-wesm87-base/CHANGELOG.md
@@ -156,4 +156,4 @@
 
 1.0.0 / 2016-04-19
 ==================
- - Initial commmit; moved content over from `eslint-config-airbnb` package.
+ - Initial commmit; moved content over from `eslint-config-wesm87` package.